mcp_tools_collection/pkg_publisher_standalone/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pkg-publisher"
version = "0.1.4"
description = "A Model Context Protocol (MCP) service for building and publishing Python packages to PyPI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "pkg-publisher contributors", email = "maintainer@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
    "requests>=2.31.0",
    "pywinpty>=2.0.0; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
pkg-publisher = "pkg_publisher.__main__:main"

[project.urls]
Homepage = "https://github.com/username/pkg-publisher"
Repository = "https://github.com/username/pkg-publisher"
Documentation = "https://github.com/username/pkg-publisher/blob/main/README.md"

[tool.setuptools.packages.find]
where = ["src"]
include = ["pkg_publisher*"]

[project.entry-points."mcp.servers"]
pkg-publisher = "pkg_publisher.__main__:main"

[tool.black]
line-length = 88

[tool.flake8]
max-line-length = 88
